Refactor: Move sorting_context class to separate files
