+++ src/limestone/sortdb_wrapper.h
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+#pragma once
 #include <boost/filesystem.hpp>
 
 #ifdef SORT_METHOD_USE_ROCKSDB
